recipe.md
To make brownies, you would have to:
Combine the melted butter, sugar, cocoa powder, vanilla, eggs, flour, baking powder, and salt. 
Spread the batter into the prepared pan. 
Decorate with walnut halves, if you would like to. 
Bake in preheated oven for 20 to 30 minutes or until a toothpick inserted in the center comes out with crumbs, which is not wet.